refactor(lib/tui): move "DownloadTracker" to TUI
as that is the only place it is used
